% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch-ucirepo.R
\name{empty_fetch_response}
\alias{empty_fetch_response}
\title{Empty Fetch Response}
\usage{
empty_fetch_response()
}
\value{
A list containing dataset metadata, dataframes, and variable info in its properties.
}
\description{
Internal function to create an empty fetch response.
}
\seealso{
\code{\link[=fetch_ucirepo]{fetch_ucirepo()}}
}
\keyword{internal}
