% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list-available-datasets.R
\name{empty_search_response}
\alias{empty_search_response}
\title{Empty Search Response}
\usage{
empty_search_response()
}
\value{
A data frame with three empty columns of 'id', 'name', and 'url'.
}
\description{
Internal function to create an empty search response.
}
\seealso{
\code{\link[=list_available_datasets]{list_available_datasets()}}
}
\keyword{internal}
