% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ucimlrepo-package.R
\docType{package}
\name{ucimlrepo-package}
\alias{ucimlrepo}
\alias{ucimlrepo-package}
\title{ucimlrepo: Import UCI ML Datasets into R}
\description{
Import datasets from the UC Irvine Machine Learning Repository into R scripts, notebooks, and 'Quarto'/'RMarkdown' documents.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r-pkg.thecoatlessprofessor.com/ucimlrepo/}
  \item \url{https://github.com/coatless-rpkg/ucimlrepo}
  \item Report bugs at \url{https://github.com/coatless-rpkg/ucimlrepo/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{james.balamuta@gmail.com} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Philip Truong \email{ucirepository@gmail.com} [copyright holder]
}

}
\keyword{internal}
